add middleware validatBody and change the code after mentors comments
