+++ controllers/contacts.js
@@ -1,7 +1,7 @@
 // const contacts = require("../models/contacts.json");
 
-const { Contact, updateFavoriteSchema } = require("../models/contact");
-const { addSchema } = require("../models/contact");
+const { Contact } = require("../models/contact");
+// const { addSchema, updateFavoriteSchema } = require("../utils/contactValidationSchema/contactValidationSchema");
 const { HttpError, ctrlWrapper } = require("../helpers");
 
 const getAll = async (req, res) => {
@@ -19,21 +19,11 @@ const getById = async (req, res, next) => {
 };
 
 const createContact = async (req, res, next) => {
-  const { name, email, phone } = req.body;
-  const { error } = addSchema.validate({ name, email, phone });
-  if (error) {
-    switch (error.details[0].context.key) {
-      case "name":
-        return res.status(400).json({ message: "missing required Name field" });
-      case "email":
-        return res.status(400).json({ message: "missing required Email field" });
-      case "phone":
-        return res.status(400).json({ message: "missing required Phone field" });
-
-      default:
-        return res.status(400).json({ message: "Something went wrong" });
-    }
-  }
+  // const { name, email, phone } = req.body;
+  // const { error } = addSchema.validate({ name, email, phone });
+  // if (error) {
+  //   throw HttpError(400, error.message);
+  // }
   const result = await Contact.create(req.body);
   res.status(201).json(result);
 
@@ -41,10 +31,10 @@ const createContact = async (req, res, next) => {
 };
 
 const updateContact = async (req, res, next) => {
-  const { error } = addSchema.validate(req.body);
-  if (error) {
-    throw HttpError({ status: 400, message: "missing fields" });
-  }
+  // const { error } = addSchema.validate(req.body);
+  // if (error) {
+  //   throw HttpError({ status: 400, message: "missing fields" });
+  // }
 
   const { contactId } = req.params;
   const result = await Contact.findByIdAndUpdate(contactId, req.body, { new: true });
@@ -55,10 +45,10 @@ const updateContact = async (req, res, next) => {
 };
 
 const updateFavorite = async (req, res, next) => {
-  const { error } = updateFavoriteSchema.validate(req.body);
-  if (error) {
-    throw HttpError({ status: 400, message: "missing fields" });
-  }
+  // const { error } = updateFavoriteSchema.validate(req.body);
+  // if (error) {
+  //   throw HttpError({ status: 400, message: "missing fields" });
+  // }
 
   const { contactId } = req.params;
   const result = await Contact.findByIdAndUpdate(contactId, req.body, { new: true });
