+++ middlewares/index.js
@@ -1,5 +1,7 @@
 const isValidId = require("./isValidId");
+const validateBody = require("./validateBody");
 
 module.exports = {
   isValidId,
+  validateBody,
 };
