+++ models/contact.js
@@ -1,5 +1,4 @@
 const { Schema, model } = require("mongoose");
-const Joi = require("joi");
 
 const contactSchema = new Schema(
   {
@@ -21,21 +20,6 @@ const contactSchema = new Schema(
   { versionKey: "false", timestamps: true }
 );
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string()
-    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net"] } })
-    .required(),
-  phone: Joi.string()
-    .pattern(/^\(\d{3}\) \d{3}-\d{4}$/)
-    .required(),
-  favorite: Joi.boolean(),
-});
-
-const updateFavoriteSchema = Joi.object({
-  favorite: Joi.boolean().required(),
-});
-
 contactSchema.post("save", (error, data, next) => {
   error.status = 400;
   next();
@@ -43,11 +27,6 @@ contactSchema.post("save", (error, data, next) => {
 
 const Contact = model("contact", contactSchema);
 
-// const schemas = {
-//   addSchema,
-// };
 module.exports = {
   Contact,
-  addSchema,
-  updateFavoriteSchema,
 };
